Collated results from Vortex scenarios - Pacioni et al. (2017)

Subset (only 3 runs) of data from Pacioni et al. (2017) used to conduct a sensitivity analysis on demographic parameters. Vortex outputs, from the project named 'Pacioni_et_al' and (Single-Factor) sensitivity test scenario 'ST_Classic' (.stdat files), were collated with collate_dat.

Format

a data.frame of 2904 observations of 68 variables.

Source

Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (2017) Predators and genetic fitness: key threatening factors for the conservation of bettong species. Pacific Conservation Biology. DOI:10.1071/PC17002

Examples

data("pac.clas")
head(pac.clas)

#>       scen.name     pop.name Year PExtant SE.PExtant. PExtinct SE.PExtinct.
#> 1 ST_Classic(1) Population 1    0       1           0        0            0
#> 2 ST_Classic(1) Population 1    1       1           0        0            0
#> 3 ST_Classic(1) Population 1    2       1           0        0            0
#> 4 ST_Classic(1) Population 1    3       1           0        0            0
#> 5 ST_Classic(1) Population 1    4       1           0        0            0
#> 6 ST_Classic(1) Population 1    5       1           0        0            0
#>    r.det r.stoch  SE.r.  SD.r. Nextant SE.Nextant. SD.Nextant.   Nall SE.Nall.
#> 1 0.1736  0.0000 0.0000 0.0000  300.00        0.00        0.00 300.00     0.00
#> 2 0.1736  0.1878 0.0089 0.0154  362.00        3.21        5.57 362.00     3.21
#> 3 0.1736  0.1806 0.0075 0.0131  433.67        4.37        7.57 433.67     4.37
#> 4 0.1736  0.1474 0.0098 0.0170  500.67        9.39       16.26 500.67     9.39
#> 5 0.1736  0.1052 0.0198 0.0344  481.33       31.49       54.54 481.33    31.49
#> 6 0.1736  0.0746 0.0782 0.1354  480.00       36.69       63.55 480.00    36.69
#>   SD.Nall.    Het SE.Het. SD.Het.  Inbr SE.Inbr. SD.Inbr. Nalleles SE.Nalleles.
#> 1     0.00 0.9983   0e+00   0e+00 0e+00    0e+00   0.0000   600.00         0.00
#> 2     5.57 0.9980   0e+00   0e+00 0e+00    0e+00   0.0000   572.00         6.51
#> 3     7.57 0.9978   0e+00   0e+00 0e+00    0e+00   0.0000   562.67         7.17
#> 4    16.26 0.9977   0e+00   0e+00 0e+00    0e+00   0.0000   549.00         5.57
#> 5    54.54 0.9975   1e-04   2e-04 8e-04    8e-04   0.0013   511.33        19.64
#> 6    63.55 0.9973   2e-04   3e-04 8e-04    8e-04   0.0014   489.33        26.34
#>   SD.Nalleles. Nlethals SE.Nlethals. SD.Nlethals. TE SE.TE. SD.TE. medTE SV1
#> 1         0.00        0            0            0  0      0      0     0 500
#> 2        11.27        0            0            0  0      0      0     0 500
#> 3        12.42        0            0            0  0      0      0     0 500
#> 4         9.64        0            0            0  0      0      0     0 500
#> 5        34.02        0            0            0  0      0      0     0 500
#> 6        45.62        0            0            0  0      0      0     0 500
#>   SE.SV1. SD.SV1. SV2 SE.SV2. SD.SV2. SV3 SE.SV3. SD.SV3. SV4 SE.SV4. SD.SV4.
#> 1       0       0   1       0       0   1       0       0   1       0       0
#> 2       0       0   1       0       0   1       0       0   1       0       0
#> 3       0       0   1       0       0   1       0       0   1       0       0
#> 4       0       0   1       0       0   1       0       0   1       0       0
#> 5       0       0   1       0       0   1       0       0   1       0       0
#> 6       0       0   1       0       0   1       0       0   1       0       0
#>   SV5 SE.SV5. SD.SV5. SV6 SE.SV6. SD.SV6. SV7 SE.SV7. SD.SV7.   PS1 SE.PS1.
#> 1 0.1       0       0  80       0       0 300       0       0 0e+00   0e+00
#> 2 0.1       0       0  80       0       0 300       0       0 0e+00   0e+00
#> 3 0.1       0       0  80       0       0 300       0       0 0e+00   0e+00
#> 4 0.1       0       0  80       0       0 300       0       0 0e+00   0e+00
#> 5 0.1       0       0  80       0       0 300       0       0 5e-04   5e-04
#> 6 0.1       0       0  80       0       0 300       0       0 5e-04   5e-04
#>   SD.PS1.    PS2 SE.PS2. SD.PS2. PS3 SE.PS3. SD.PS3.    PS4 SE.PS4. SD.PS4.
#> 1   0e+00 0.1998  0.0012  0.0022 4.4       0       0 0.4731  0.0849  0.1471
#> 2   0e+00 0.1996  0.0010  0.0017 4.4       0       0 0.6407  0.1499  0.2597
#> 3   0e+00 0.1985  0.0017  0.0030 4.4       0       0 0.2543  0.0826  0.1430
#> 4   0e+00 0.1982  0.0017  0.0029 4.4       0       0 0.3229  0.1573  0.2725
#> 5   8e-04 0.2002  0.0018  0.0031 4.4       0       0 0.4483  0.1931  0.3344
#> 6   9e-04 0.2003  0.0016  0.0028 4.4       0       0 0.2188  0.0401  0.0694
#>   SD.PExtant. SD.PExtinct.
#> 1           0            0
#> 2           0            0
#> 3           0            0
#> 4           0            0
#> 5           0            0
#> 6           0            0