Results of pairwise comparisons of simulation scenarios

Results of pairwise comparisons of simulation scenarios included in the sensitivity test scenario 'ST_Classic' using pairwise.

Format

A named list of 12 elements. See documentation for details.

Source

Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (2017) Predators and genetic fitness: key threatening factors for the conservation of bettong species. Pacific Conservation Biology. DOI:10.1071/PC17002

Examples

data("pac.clas.pairw")
head(pac.clas.pairw)

#> $coef.table
#>          Scenario   Population    SC_Nall60      SC_Het60   SC_Nall120
#> 1   ST_Classic(1) Population 1 -0.753106740 -0.0090488639 -0.745396246
#> 2   ST_Classic(2) Population 1 -0.553980065 -0.0032173738 -0.545881126
#> 3   ST_Classic(3) Population 1 -0.041797577 -0.0009048864  0.047095933
#> 4   ST_Classic(4) Population 1 -0.126031748 -0.0017092298  0.054223149
#> 5   ST_Classic(5) Population 1 -0.073264365 -0.0029157450 -0.101840459
#> 6   ST_Classic(6) Population 1 -0.087463148 -0.0063342047 -0.043623566
#> 7   ST_Classic(7) Population 1 -0.207361292 -0.0098532073 -0.106882169
#> 8   ST_Classic(8) Population 1 -0.063098172  0.0003016288  0.075599583
#> 9   ST_Classic(9) Population 1 -0.049543247 -0.0001005429 -0.110703858
#> 10 ST_Classic(10) Population 1 -0.140714635 -0.0014076010  0.034760167
#> 11 ST_Classic(11) Population 1 -0.192029705 -0.0026141162 -0.055615224
#> 12 ST_Classic(12) Population 1 -0.975150920 -0.0292579932 -0.995135558
#> 13 ST_Classic(13) Population 1 -0.001776663 -0.0002010859  0.049880083
#> 14 ST_Classic(14) Population 1 -0.100693722  0.0001005429  0.030933264
#> 15 ST_Classic(15) Population 1 -0.009362580  0.0003016288  0.025547445
#> 16 ST_Classic(16) Population 1 -0.039212459  0.0002010859 -0.035974974
#> 17 ST_Classic(17) Population 1 -0.016459551  0.0001005429  0.096105318
#> 18 ST_Classic(18) Population 1 -0.011298997  0.0003016288  0.008342023
#> 19 ST_Classic(19) Population 1 -0.033727556 -0.0002010859  0.000000000
#> 20 ST_Classic(20) Population 1 -0.007745671  0.0000000000 -0.035453597
#> 21 ST_Classic(21) Population 1 -0.083750067 -0.0003016288  0.063779979
#> 22 ST_Classic(22) Population 1 -0.023077258 -0.0004021717  0.114353493
#> 23 ST_Classic(23) Population 1 -0.061965367 -0.0065352906  0.062737226
#>        SC_Het120
#> 1  -0.0197640415
#> 2  -0.0075627710
#> 3  -0.0012100434
#> 4  -0.0028234345
#> 5  -0.0038318040
#> 6  -0.0063527276
#> 7  -0.0129071292
#> 8   0.0000000000
#> 9  -0.0003025108
#> 10 -0.0008066956
#> 11 -0.0028234345
#> 12 -1.0000000000
#> 13 -0.0006050217
#> 14 -0.0004033478
#> 15  0.0001008369
#> 16  0.0003025108
#> 17  0.0002016739
#> 18 -0.0006050217
#> 19 -0.0001008369
#> 20 -0.0004033478
#> 21 -0.0003025108
#> 22 -0.0003025108
#> 23 -0.0063527276
#> 
#> $SSMD.table
#>          Scenario   Population  SSMD_Nall60  SSMD_Het60 SSMD_Nall120
#> 1   ST_Classic(1) Population 1 -10.24389822 -17.6504522  -8.26099855
#> 2   ST_Classic(2) Population 1  -7.19955038  -4.5254834  -5.59253479
#> 3   ST_Classic(3) Population 1  -0.55460296  -4.0249224   0.41904660
#> 4   ST_Classic(4) Population 1  -1.28749773  -7.6026311   0.58029707
#> 5   ST_Classic(5) Population 1  -0.84614075  -5.6873679  -0.66510998
#> 6   ST_Classic(6) Population 1  -1.06695417  -3.6994874  -0.41449553
#> 7   ST_Classic(7) Population 1  -1.57586353  -4.0797934  -0.95694624
#> 8   ST_Classic(8) Population 1  -0.75908302   1.3416408   0.83068382
#> 9   ST_Classic(9) Population 1  -0.51959543  -0.4472136  -1.18040589
#> 10 ST_Classic(10) Population 1  -1.34746470  -4.4271887   0.36870131
#> 11 ST_Classic(11) Population 1  -1.55231922 -18.3847763  -0.54150939
#> 12 ST_Classic(12) Population 1 -13.33285833  -5.4895890 -10.99812985
#> 13 ST_Classic(13) Population 1  -0.01986106  -0.6324555   0.48193240
#> 14 ST_Classic(14) Population 1  -0.98084413   0.3162278   0.18928233
#> 15 ST_Classic(15) Population 1  -0.11745726   3.0000000   0.26307009
#> 16 ST_Classic(16) Population 1  -0.52733561   0.6324555  -0.31656651
#> 17 ST_Classic(17) Population 1  -0.21575550   0.3162278   1.06177944
#> 18 ST_Classic(18) Population 1  -0.14765357   0.9486833   0.07427443
#> 19 ST_Classic(19) Population 1  -0.32536520  -0.3922323   0.00000000
#> 20 ST_Classic(20) Population 1  -0.09260888   0.0000000  -0.22067003
#> 21 ST_Classic(21) Population 1  -1.12142164  -0.5883484   0.44096584
#> 22 ST_Classic(22) Population 1  -0.30079575  -4.0000000   1.17067958
#> 23 ST_Classic(23) Population 1  -0.76444287  -9.1923882   0.63823134
#>      SSMD_Het120
#> 1     -7.4508020
#> 2     -2.7477874
#> 3     -2.6832816
#> 4     -3.4729726
#> 5     -5.9346030
#> 6     -4.9805873
#> 7    -10.9357780
#> 8      0.0000000
#> 9     -0.5303301
#> 10    -1.4142136
#> 11    -2.0586009
#> 12 -2479.2500000
#> 13    -0.6708204
#> 14    -0.6246950
#> 15     0.2000000
#> 16     0.3721042
#> 17     0.3535534
#> 18    -0.5570860
#> 19    -0.1386750
#> 20    -0.3713907
#> 21    -0.6708204
#> 22    -0.3354102
#> 23    -7.0436141
#> 
#> $SSMD.table.pvalues
#>          Scenario   Population  SSMD_Nall60   SSMD_Het60 SSMD_Nall120
#> 1   ST_Classic(1) Population 1 6.302641e-25 5.047239e-70 7.222942e-17
#> 2   ST_Classic(2) Population 1 3.020572e-13 3.012881e-06 1.118891e-08
#> 3   ST_Classic(3) Population 1 2.895831e-01 2.849706e-05 3.375910e-01
#> 4   ST_Classic(4) Population 1 9.896043e-02 1.450849e-14 2.808572e-01
#> 5   ST_Classic(5) Population 1 1.987371e-01 6.450616e-09 2.529901e-01
#> 6   ST_Classic(6) Population 1 1.429963e-01 1.080177e-04 3.392556e-01
#> 7   ST_Classic(7) Population 1 5.752863e-02 2.253787e-05 1.692972e-01
#> 8   ST_Classic(8) Population 1 2.239014e-01 8.985625e-02 2.030761e-01
#> 9   ST_Classic(9) Population 1 3.016728e-01 3.273604e-01 1.189194e-01
#> 10 ST_Classic(10) Population 1 8.891531e-02 4.773460e-06 3.561752e-01
#> 11 ST_Classic(11) Population 1 6.029293e-02 8.697787e-76 2.940783e-01
#> 12 ST_Classic(12) Population 1 7.453731e-41 2.014351e-08 1.950695e-28
#> 13 ST_Classic(13) Population 1 4.920771e-01 2.635446e-01 3.149270e-01
#> 14 ST_Classic(14) Population 1 1.633348e-01 3.759148e-01 4.249358e-01
#> 15 ST_Classic(15) Population 1 4.532489e-01 1.349898e-03 3.962483e-01
#> 16 ST_Classic(16) Population 1 2.989803e-01 2.635446e-01 3.757863e-01
#> 17 ST_Classic(17) Population 1 4.145892e-01 3.759148e-01 1.441679e-01
#> 18 ST_Classic(18) Population 1 4.413081e-01 1.713909e-01 4.703960e-01
#> 19 ST_Classic(19) Population 1 3.724523e-01 3.474433e-01 5.000000e-01
#> 20 ST_Classic(20) Population 1 4.631071e-01 5.000000e-01 4.126747e-01
#> 21 ST_Classic(21) Population 1 1.310542e-01 2.781492e-01 3.296189e-01
#> 22 ST_Classic(22) Population 1 3.817851e-01 3.167124e-05 1.208638e-01
#> 23 ST_Classic(23) Population 1 2.223017e-01 1.921074e-20 2.616615e-01
#>     SSMD_Het120
#> 1  4.638725e-14
#> 2  2.999945e-03
#> 3  3.645179e-03
#> 4  2.573640e-04
#> 5  1.472787e-09
#> 6  3.169580e-07
#> 7  3.886815e-28
#> 8  5.000000e-01
#> 9  2.979415e-01
#> 10 7.864960e-02
#> 11 1.976624e-02
#> 12 0.000000e+00
#> 13 2.511675e-01
#> 14 2.660856e-01
#> 15 4.207403e-01
#> 16 3.549076e-01
#> 17 3.618368e-01
#> 18 2.887343e-01
#> 19 4.448535e-01
#> 20 3.551733e-01
#> 21 2.511675e-01
#> 22 3.686578e-01
#> 23 9.365822e-13
#> 
#> $ranks.SC
#>       Population       Scenario SC_Nall60 SC_Het60 SC_Nall120 SC_Het120
#>  1: Population 1  ST_Classic(1)         2      3.0          2       2.0
#>  2: Population 1  ST_Classic(2)         3      6.0          3       4.0
#>  3: Population 1  ST_Classic(3)        15     11.0         15      10.0
#>  4: Population 1  ST_Classic(4)         7      9.0         13       8.5
#>  5: Population 1  ST_Classic(5)        11      7.0          7       7.0
#>  6: Population 1  ST_Classic(6)         9      5.0         16       5.5
#>  7: Population 1  ST_Classic(7)         4      2.0          6       3.0
#>  8: Population 1  ST_Classic(8)        12     15.0          9      23.0
#>  9: Population 1  ST_Classic(9)        14     21.0          5      17.0
#> 10: Population 1 ST_Classic(10)         6     10.0         19      11.0
#> 11: Population 1 ST_Classic(11)         5      8.0         12       8.5
#> 12: Population 1 ST_Classic(12)         1      1.0          1       1.0
#> 13: Population 1 ST_Classic(13)        23     17.5         14      12.5
#> 14: Population 1 ST_Classic(14)         8     21.0         20      14.5
#> 15: Population 1 ST_Classic(15)        21     15.0         21      21.5
#> 16: Population 1 ST_Classic(16)        16     19.0         17      19.0
#> 17: Population 1 ST_Classic(17)        19     21.0          8      20.0
#> 18: Population 1 ST_Classic(18)        20     15.0         22      12.5
#> 19: Population 1 ST_Classic(19)        17     17.5         23      21.5
#> 20: Population 1 ST_Classic(20)        22     23.0         18      14.5
#> 21: Population 1 ST_Classic(21)        10     13.0         10      17.0
#> 22: Population 1 ST_Classic(22)        18     12.0          4      17.0
#> 23: Population 1 ST_Classic(23)        13      4.0         11       5.5
#>       Population       Scenario SC_Nall60 SC_Het60 SC_Nall120 SC_Het120
#> 
#> $ranks.SSMD
#>       Population       Scenario SSMD_Nall60 SSMD_Het60 SSMD_Nall120 SSMD_Het120
#>  1: Population 1  ST_Classic(1)           2        2.0            2           3
#>  2: Population 1  ST_Classic(2)           3        7.0            3           8
#>  3: Population 1  ST_Classic(3)          14       10.0           15           9
#>  4: Population 1  ST_Classic(4)           7        4.0           11           7
#>  5: Population 1  ST_Classic(5)          11        5.0            9           5
#>  6: Population 1  ST_Classic(6)           9       12.0           16           6
#>  7: Population 1  ST_Classic(7)           4        9.0            7           2
#>  8: Population 1  ST_Classic(8)          13       14.0            8          23
#>  9: Population 1  ST_Classic(9)          16       19.0            4          16
#> 10: Population 1 ST_Classic(10)           6        8.0           17          11
#> 11: Population 1 ST_Classic(11)           5        1.0           12          10
#> 12: Population 1 ST_Classic(12)           1        6.0            1           1
#> 13: Population 1 ST_Classic(13)          23       16.0           13          13
#> 14: Population 1 ST_Classic(14)          10       21.5           21          14
#> 15: Population 1 ST_Classic(15)          21       13.0           19          21
#> 16: Population 1 ST_Classic(16)          15       17.0           18          17
#> 17: Population 1 ST_Classic(17)          19       21.5            6          19
#> 18: Population 1 ST_Classic(18)          20       15.0           22          15
#> 19: Population 1 ST_Classic(19)          17       20.0           23          22
#> 20: Population 1 ST_Classic(20)          22       23.0           20          18
#> 21: Population 1 ST_Classic(21)           8       18.0           14          12
#> 22: Population 1 ST_Classic(22)          18       11.0            5          20
#> 23: Population 1 ST_Classic(23)          12        3.0           10           4
#>       Population       Scenario SSMD_Nall60 SSMD_Het60 SSMD_Nall120 SSMD_Het120
#> 
#> $Kendall
#> $Kendall$SC
#>  [1] "Rank comparison of sensitivity coefficients "
#>  [2] "$`Population 1`"                             
#>  [3] " Kendall's coefficient of concordance Wt"    
#>  [4] ""                                            
#>  [5] " Subjects = 23 "                             
#>  [6] "   Raters = 4 "                              
#>  [7] "       Wt = 0.712 "                          
#>  [8] ""                                            
#>  [9] "Chisq(22) = 62.6 "                           
#> [10] "  p-value = 9.1e-06 "                        
#> [11] ""                                            
#> 
#> $Kendall$SSMD
#>  [1] "Rank comparison of SSMD "                
#>  [2] "$`Population 1`"                         
#>  [3] " Kendall's coefficient of concordance Wt"
#>  [4] ""                                        
#>  [5] " Subjects = 23 "                         
#>  [6] "   Raters = 4 "                          
#>  [7] "       Wt = 0.685 "                      
#>  [8] ""                                        
#>  [9] "Chisq(22) = 60.3 "                       
#> [10] "  p-value = 2.05e-05 "                   
#> [11] ""                                        
#> 
#>