Collated results from Vortex scenarios - Pacioni et al. (2017)

Data from Pacioni et al. (2017) used to conduct a sensitivity analysis on demographic parameters. Vortex outputs, from the project named 'Pacioni_et_al' and (Latin Hypercube Sampling) sensitivity test scenario 'ST_LHS' (.stdat files), were collated with collate_dat.

Format

A data.frame of 6171 observations of 68 variables.

Source

Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (2017) Predators and genetic fitness: key threatening factors for the conservation of bettong species. Pacific Conservation Biology. DOI:10.1071/PC17002

Examples

data("pac.lhs")
head(pac.lhs)

#>   scen.name     pop.name Year PExtant SE.PExtant. PExtinct SE.PExtinct.  r.det
#> 1 ST_LHS(1) Population 1    0       1           0        0            0 -0.066
#> 2 ST_LHS(1) Population 1    1       1           0        0            0 -0.066
#> 3 ST_LHS(1) Population 1    2       1           0        0            0 -0.066
#> 4 ST_LHS(1) Population 1    3       1           0        0            0 -0.066
#> 5 ST_LHS(1) Population 1    4       1           0        0            0 -0.066
#> 6 ST_LHS(1) Population 1    5       1           0        0            0 -0.066
#>   r.stoch  SE.r.  SD.r. Nextant SE.Nextant. SD.Nextant.   Nall SE.Nall.
#> 1  0.0000 0.0000 0.0000  108.00        0.00        0.00 108.00     0.00
#> 2 -0.0359 0.0675 0.1169  104.67        6.96       12.06 104.67     6.96
#> 3 -0.0915 0.0178 0.0309   95.33        4.91        8.50  95.33     4.91
#> 4 -0.0426 0.0413 0.0715   91.33        4.41        7.64  91.33     4.41
#> 5 -0.0900 0.0234 0.0406   83.33        2.19        3.79  83.33     2.19
#> 6 -0.1325 0.0112 0.0193   73.00        2.08        3.61  73.00     2.08
#>   SD.Nall.    Het SE.Het. SD.Het. Inbr SE.Inbr. SD.Inbr. Nalleles SE.Nalleles.
#> 1     0.00 0.9954   0e+00   0e+00    0        0        0   216.00         0.00
#> 2    12.06 0.9935   4e-04   6e-04    0        0        0   173.33         9.91
#> 3     8.50 0.9917   4e-04   8e-04    0        0        0   141.67         7.13
#> 4     7.64 0.9900   4e-04   7e-04    0        0        0   124.00         4.58
#> 5     3.79 0.9889   2e-04   3e-04    0        0        0   110.00         2.08
#> 6     3.61 0.9868   2e-04   3e-04    0        0        0    94.67         1.67
#>   SD.Nalleles. Nlethals SE.Nlethals. SD.Nlethals. TE SE.TE. SD.TE. medTE
#> 1         0.00        0            0            0  0      0      0     0
#> 2        17.16        0            0            0  0      0      0     0
#> 3        12.34        0            0            0  0      0      0     0
#> 4         7.94        0            0            0  0      0      0     0
#> 5         3.61        0            0            0  0      0      0     0
#> 6         2.89        0            0            0  0      0      0     0
#>        SV1 SE.SV1. SD.SV1.    SV2 SE.SV2. SD.SV2.    SV3 SE.SV3. SD.SV3.    SV4
#> 1 1540.816       0       0 4.5714       0       0 3.8571       0       0 5.2857
#> 2 1540.816       0       0 4.5714       0       0 3.8571       0       0 5.2857
#> 3 1540.816       0       0 4.5714       0       0 3.8571       0       0 5.2857
#> 4 1540.816       0       0 4.5714       0       0 3.8571       0       0 5.2857
#> 5 1540.816       0       0 4.5714       0       0 3.8571       0       0 5.2857
#> 6 1540.816       0       0 4.5714       0       0 3.8571       0       0 5.2857
#>   SE.SV4. SD.SV4.   SV5 SE.SV5. SD.SV5.     SV6 SE.SV6. SD.SV6.      SV7
#> 1       0       0 0.198       0       0 90.2041       0       0 108.1633
#> 2       0       0 0.198       0       0 90.2041       0       0 108.1633
#> 3       0       0 0.198       0       0 90.2041       0       0 108.1633
#> 4       0       0 0.198       0       0 90.2041       0       0 108.1633
#> 5       0       0 0.198       0       0 90.2041       0       0 108.1633
#> 6       0       0 0.198       0       0 90.2041       0       0 108.1633
#>   SE.SV7. SD.SV7. PS1 SE.PS1. SD.PS1.    PS2 SE.PS2. SD.PS2.     PS3 SE.PS3.
#> 1       0       0   0       0       0 0.2051  0.0029  0.0050 23.2571       0
#> 2       0       0   0       0       0 0.1943  0.0020  0.0034 23.2571       0
#> 3       0       0   0       0       0 0.1968  0.0033  0.0057 23.2571       0
#> 4       0       0   0       0       0 0.1958  0.0020  0.0034 23.2571       0
#> 5       0       0   0       0       0 0.1930  0.0038  0.0066 23.2571       0
#> 6       0       0   0       0       0 0.1942  0.0026  0.0044 23.2571       0
#>   SD.PS3.    PS4 SE.PS4. SD.PS4. SD.PExtant. SD.PExtinct.
#> 1       0 0.4877  0.2391  0.4142           0            0
#> 2       0 0.7526  0.0681  0.1179           0            0
#> 3       0 0.3393  0.1468  0.2542           0            0
#> 4       0 0.6252  0.2558  0.4430           0            0
#> 5       0 0.3964  0.1834  0.3176           0            0
#> 6       0 0.6452  0.2399  0.4155           0            0