Collated results from Vortex scenarios - Pacioni et al. (2017)

Data from Pacioni et al. (2017) used to conduct a sensitivity analysis on demographic parameters. Vortex outputs, from the project named 'Pacioni_et_al' and (Latin Hypercube Sampling) sensitivity test scenario 'ST_LHS' (.run files), were collated with collate_run.

Format

A named list of two data.frames: run (153 obs, 7 var), lrun (153 obs, 8 var).

Source

Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (2017) Predators and genetic fitness: key threatening factors for the conservation of bettong species. Pacific Conservation Biology. DOI:10.1071/PC17002

Examples

data("pac.run.lhs")
head(pac.run.lhs)

#> $run
#>         Scenario Iteration YrExt_Population1 N_Population1 GeneDiv_Population1
#> 1      ST_LHS(1)         1                 8             0                  NA
#> 2      ST_LHS(1)         2                12             0                  NA
#> 3      ST_LHS(1)         3                13             0                  NA
#> 4      ST_LHS(2)         1                NA          1384              0.9605
#> 5      ST_LHS(2)         2                NA          1514              0.9372
#> 6      ST_LHS(2)         3                NA           605              0.8284
#> 7      ST_LHS(3)         1                NA          1364              0.9883
#> 8      ST_LHS(3)         2                NA          1366              0.9839
#> 9      ST_LHS(3)         3                NA          1108              0.9823
#> 10     ST_LHS(4)         1                 7             0                  NA
#> 11     ST_LHS(4)         2                 5             0                  NA
#> 12     ST_LHS(4)         3                 4             0                  NA
#> 13     ST_LHS(5)         1                19             0                  NA
#> 14     ST_LHS(5)         2                19             0                  NA
#> 15     ST_LHS(5)         3                32             0                  NA
#> 16     ST_LHS(6)         1                11             0                  NA
#> 17     ST_LHS(6)         2                13             0                  NA
#> 18     ST_LHS(6)         3                15             0                  NA
#> 19     ST_LHS(7)         1                79            16                  NA
#> 20     ST_LHS(7)         2                54            13                  NA
#> 21     ST_LHS(7)         3                83             3                  NA
#> 22     ST_LHS(8)         1                70            18                  NA
#> 23     ST_LHS(8)         2                24             0                  NA
#> 24     ST_LHS(8)         3                23             0                  NA
#> 25     ST_LHS(9)         1                NA           382              0.9339
#> 26     ST_LHS(9)         2                NA           722              0.9538
#> 27     ST_LHS(9)         3                NA           386              0.9679
#> 28    ST_LHS(10)         1                32             0                  NA
#> 29    ST_LHS(10)         2                34             0                  NA
#> 30    ST_LHS(10)         3                99            15                  NA
#> 31    ST_LHS(11)         1                NA           785              0.9674
#> 32    ST_LHS(11)         2                NA           814              0.9658
#> 33    ST_LHS(11)         3                NA           809              0.9710
#> 34    ST_LHS(12)         1                NA          1253              0.9864
#> 35    ST_LHS(12)         2                NA          1330              0.9845
#> 36    ST_LHS(12)         3                NA          1290              0.9861
#> 37    ST_LHS(13)         1                33             0                  NA
#> 38    ST_LHS(13)         2                36             0                  NA
#> 39    ST_LHS(13)         3                24             0                  NA
#> 40    ST_LHS(14)         1                NA           557              0.9709
#> 41    ST_LHS(14)         2                NA           588              0.9736
#> 42    ST_LHS(14)         3                NA           527              0.9732
#> 43    ST_LHS(15)         1                NA           385              0.9575
#> 44    ST_LHS(15)         2                NA           552              0.9549
#> 45    ST_LHS(15)         3                NA           580              0.9635
#> 46    ST_LHS(16)         1                11             0                  NA
#> 47    ST_LHS(16)         2                 8             0                  NA
#> 48    ST_LHS(16)         3                11             0                  NA
#> 49    ST_LHS(17)         1                NA          1166              0.9756
#> 50    ST_LHS(17)         2                NA          1057              0.9755
#> 51    ST_LHS(17)         3                NA          1098              0.9763
#> 52    ST_LHS(18)         1                12             0                  NA
#> 53    ST_LHS(18)         2                 4             0                  NA
#> 54    ST_LHS(18)         3                 6             0                  NA
#> 55    ST_LHS(19)         1                 9             0                  NA
#> 56    ST_LHS(19)         2                 9             0                  NA
#> 57    ST_LHS(19)         3                10             0                  NA
#> 58    ST_LHS(20)         1                 9             0                  NA
#> 59    ST_LHS(20)         2                 9             0                  NA
#> 60    ST_LHS(20)         3                12             0                  NA
#> 61    ST_LHS(21)         1                NA          1106              0.9642
#> 62    ST_LHS(21)         2                NA          1271              0.9636
#> 63    ST_LHS(21)         3                NA          1544              0.9632
#> 64    ST_LHS(22)         1                NA           538              0.9716
#> 65    ST_LHS(22)         2                NA           587              0.9713
#> 66    ST_LHS(22)         3                NA           541              0.9664
#> 67    ST_LHS(23)         1                36             0                  NA
#> 68    ST_LHS(23)         2                19             0                  NA
#> 69    ST_LHS(23)         3                30             0                  NA
#> 70    ST_LHS(24)         1                NA            88              0.8917
#> 71    ST_LHS(24)         2               101            12                  NA
#> 72    ST_LHS(24)         3                NA           232              0.9211
#> 73    ST_LHS(25)         1                NA          1087              0.9744
#> 74    ST_LHS(25)         2                NA          1126              0.9703
#> 75    ST_LHS(25)         3                NA          1218              0.9742
#> 76    ST_LHS(26)         1                NA           438              0.9634
#> 77    ST_LHS(26)         2                NA           390              0.9580
#> 78    ST_LHS(26)         3                NA           418              0.9211
#> 79    ST_LHS(27)         1                NA           465              0.9588
#> 80    ST_LHS(27)         2                NA           465              0.9734
#> 81    ST_LHS(27)         3                NA           437              0.9556
#> 82    ST_LHS(28)         1                NA           464              0.9707
#> 83    ST_LHS(28)         2                NA           244              0.9585
#> 84    ST_LHS(28)         3                NA           353              0.9744
#> 85    ST_LHS(29)         1                NA           408              0.9528
#> 86    ST_LHS(29)         2                NA           468              0.9499
#> 87    ST_LHS(29)         3                NA           419              0.9727
#> 88    ST_LHS(30)         1                NA           461              0.9291
#> 89    ST_LHS(30)         2                NA           291              0.8715
#> 90    ST_LHS(30)         3                NA            79              0.9101
#> 91    ST_LHS(31)         1                NA           930              0.9830
#> 92    ST_LHS(31)         2                NA          1010              0.9845
#> 93    ST_LHS(31)         3                NA           764              0.9861
#> 94    ST_LHS(32)         1                NA           663              0.9808
#> 95    ST_LHS(32)         2                NA           580              0.9728
#> 96    ST_LHS(32)         3                NA           678              0.9803
#> 97    ST_LHS(33)         1                NA           781              0.9715
#> 98    ST_LHS(33)         2                NA           796              0.9645
#> 99    ST_LHS(33)         3                NA           847              0.9447
#> 100   ST_LHS(34)         1                NA           476              0.9742
#> 101   ST_LHS(34)         2                NA           577              0.9751
#> 102   ST_LHS(34)         3                NA           603              0.9792
#> 103   ST_LHS(35)         1                NA           355              0.9282
#> 104   ST_LHS(35)         2                NA           187              0.9366
#> 105   ST_LHS(35)         3                NA           879              0.9445
#> 106   ST_LHS(36)         1                12             0                  NA
#> 107   ST_LHS(36)         2                18             0                  NA
#> 108   ST_LHS(36)         3                13             0                  NA
#> 109   ST_LHS(37)         1                NA           982              0.9855
#> 110   ST_LHS(37)         2                NA           996              0.9838
#> 111   ST_LHS(37)         3                NA           884              0.9837
#> 112   ST_LHS(38)         1                15             0                  NA
#> 113   ST_LHS(38)         2                24             0                  NA
#> 114   ST_LHS(38)         3                21             0                  NA
#> 115   ST_LHS(39)         1                 7             0                  NA
#> 116   ST_LHS(39)         2                10             0                  NA
#> 117   ST_LHS(39)         3                10             0                  NA
#> 118   ST_LHS(40)         1                NA          1466              0.9883
#> 119   ST_LHS(40)         2                NA          1429              0.9880
#> 120   ST_LHS(40)         3                NA          1464              0.9882
#> 121   ST_LHS(41)         1                29             0                  NA
#> 122   ST_LHS(41)         2                24             0                  NA
#> 123   ST_LHS(41)         3                22             0                  NA
#> 124   ST_LHS(42)         1                NA           755              0.9612
#> 125   ST_LHS(42)         2                NA           731              0.9546
#> 126   ST_LHS(42)         3                NA           776              0.9663
#> 127   ST_LHS(43)         1                31             0                  NA
#> 128   ST_LHS(43)         2                55             0                  NA
#> 129   ST_LHS(43)         3                20             0                  NA
#> 130   ST_LHS(44)         1                33             0                  NA
#> 131   ST_LHS(44)         2                39             0                  NA
#> 132   ST_LHS(44)         3                35            10                  NA
#> 133   ST_LHS(45)         1                48             0                  NA
#> 134   ST_LHS(45)         2                50             0                  NA
#> 135   ST_LHS(45)         3                30             0                  NA
#> 136   ST_LHS(46)         1                NA          1243              0.9709
#> 137   ST_LHS(46)         2                NA          1136              0.9795
#> 138   ST_LHS(46)         3                NA          1216              0.9774
#> 139   ST_LHS(47)         1               109            23                  NA
#> 140   ST_LHS(47)         2                NA            52              0.8600
#> 141   ST_LHS(47)         3                NA           379              0.9333
#> 142   ST_LHS(48)         1                NA           213              0.9194
#> 143   ST_LHS(48)         2                NA           183              0.8876
#> 144   ST_LHS(48)         3                NA            81              0.8112
#> 145   ST_LHS(49)         1                NA           682              0.9502
#> 146   ST_LHS(49)         2                NA           681              0.9550
#> 147   ST_LHS(49)         3                NA           690              0.9524
#> 148   ST_LHS(50)         1                NA          1197              0.9884
#> 149   ST_LHS(50)         2                NA          1036              0.9887
#> 150   ST_LHS(50)         3                NA           993              0.9875
#> 151 ST_LHS(Base)         1                 1             0                  NA
#> 152 ST_LHS(Base)         2                 1             0                  NA
#> 153 ST_LHS(Base)         3                 1             0                  NA
#>     Inbreed_Population1 Alleles_Population1
#> 1                    NA                  NA
#> 2                    NA                  NA
#> 3                    NA                  NA
#> 4                0.0376                  51
#> 5                0.0707                  31
#> 6                0.1835                  12
#> 7                0.0110                 146
#> 8                0.0205                 113
#> 9                0.0162                 119
#> 10                   NA                  NA
#> 11                   NA                  NA
#> 12                   NA                  NA
#> 13                   NA                  NA
#> 14                   NA                  NA
#> 15                   NA                  NA
#> 16                   NA                  NA
#> 17                   NA                  NA
#> 18                   NA                  NA
#> 19                   NA                  NA
#> 20                   NA                  NA
#> 21                   NA                  NA
#> 22                   NA                  NA
#> 23                   NA                  NA
#> 24                   NA                  NA
#> 25               0.0628                  28
#> 26               0.0402                  42
#> 27               0.0181                  46
#> 28                   NA                  NA
#> 29                   NA                  NA
#> 30                   NA                  NA
#> 31               0.0217                  62
#> 32               0.0319                  64
#> 33               0.0334                  71
#> 34               0.0160                 127
#> 35               0.0180                 125
#> 36               0.0116                 124
#> 37                   NA                  NA
#> 38                   NA                  NA
#> 39                   NA                  NA
#> 40               0.0215                  67
#> 41               0.0221                  68
#> 42               0.0285                  63
#> 43               0.0286                  39
#> 44               0.0489                  49
#> 45               0.0362                  50
#> 46                   NA                  NA
#> 47                   NA                  NA
#> 48                   NA                  NA
#> 49               0.0232                  77
#> 50               0.0237                  74
#> 51               0.0246                  80
#> 52                   NA                  NA
#> 53                   NA                  NA
#> 54                   NA                  NA
#> 55                   NA                  NA
#> 56                   NA                  NA
#> 57                   NA                  NA
#> 58                   NA                  NA
#> 59                   NA                  NA
#> 60                   NA                  NA
#> 61               0.0407                  48
#> 62               0.0393                  47
#> 63               0.0298                  59
#> 64               0.0353                  65
#> 65               0.0187                  55
#> 66               0.0314                  50
#> 67                   NA                  NA
#> 68                   NA                  NA
#> 69                   NA                  NA
#> 70               0.0909                  11
#> 71                   NA                  NA
#> 72               0.0776                  23
#> 73               0.0313                  69
#> 74               0.0284                  68
#> 75               0.0263                  62
#> 76               0.0137                  42
#> 77               0.0359                  41
#> 78               0.0622                  29
#> 79               0.0409                  42
#> 80               0.0129                  60
#> 81               0.0526                  47
#> 82               0.0302                  55
#> 83               0.0492                  38
#> 84               0.0283                  61
#> 85               0.0515                  47
#> 86               0.0769                  44
#> 87               0.0167                  61
#> 88               0.0651                  23
#> 89               0.1375                  14
#> 90               0.1013                  16
#> 91               0.0183                 112
#> 92               0.0149                 118
#> 93               0.0196                 124
#> 94               0.0226                  87
#> 95               0.0276                  74
#> 96               0.0206                  88
#> 97               0.0320                  58
#> 98               0.0226                  46
#> 99               0.0626                  41
#> 100              0.0336                  74
#> 101              0.0191                  73
#> 102              0.0116                  81
#> 103              0.0620                  34
#> 104              0.0802                  24
#> 105              0.0478                  32
#> 106                  NA                  NA
#> 107                  NA                  NA
#> 108                  NA                  NA
#> 109              0.0102                 118
#> 110              0.0050                 116
#> 111              0.0147                 121
#> 112                  NA                  NA
#> 113                  NA                  NA
#> 114                  NA                  NA
#> 115                  NA                  NA
#> 116                  NA                  NA
#> 117                  NA                  NA
#> 118              0.0109                 149
#> 119              0.0140                 154
#> 120              0.0075                 160
#> 121                  NA                  NA
#> 122                  NA                  NA
#> 123                  NA                  NA
#> 124              0.0477                  49
#> 125              0.0451                  34
#> 126              0.0245                  61
#> 127                  NA                  NA
#> 128                  NA                  NA
#> 129                  NA                  NA
#> 130                  NA                  NA
#> 131                  NA                  NA
#> 132                  NA                  NA
#> 133                  NA                  NA
#> 134                  NA                  NA
#> 135                  NA                  NA
#> 136              0.0241                  61
#> 137              0.0194                  89
#> 138              0.0214                  80
#> 139                  NA                  NA
#> 140              0.0962                  10
#> 141              0.0580                  26
#> 142              0.0704                  21
#> 143              0.1475                  21
#> 144              0.2222                  13
#> 145              0.0469                  38
#> 146              0.0426                  40
#> 147              0.0348                  39
#> 148              0.0109                 157
#> 149              0.0068                 156
#> 150              0.0111                 134
#> 151                  NA                  NA
#> 152                  NA                  NA
#> 153                  NA                  NA
#> 
#> $lrun
#>      Population     Scenario Iteration YrExt    N GeneDiv Inbreed Alleles
#> 1   Population1    ST_LHS(1)         1     8    0      NA      NA      NA
#> 2   Population1    ST_LHS(1)         2    12    0      NA      NA      NA
#> 3   Population1    ST_LHS(1)         3    13    0      NA      NA      NA
#> 4   Population1    ST_LHS(2)         1    NA 1384  0.9605  0.0376      51
#> 5   Population1    ST_LHS(2)         2    NA 1514  0.9372  0.0707      31
#> 6   Population1    ST_LHS(2)         3    NA  605  0.8284  0.1835      12
#> 7   Population1    ST_LHS(3)         1    NA 1364  0.9883  0.0110     146
#> 8   Population1    ST_LHS(3)         2    NA 1366  0.9839  0.0205     113
#> 9   Population1    ST_LHS(3)         3    NA 1108  0.9823  0.0162     119
#> 10  Population1    ST_LHS(4)         1     7    0      NA      NA      NA
#> 11  Population1    ST_LHS(4)         2     5    0      NA      NA      NA
#> 12  Population1    ST_LHS(4)         3     4    0      NA      NA      NA
#> 13  Population1    ST_LHS(5)         1    19    0      NA      NA      NA
#> 14  Population1    ST_LHS(5)         2    19    0      NA      NA      NA
#> 15  Population1    ST_LHS(5)         3    32    0      NA      NA      NA
#> 16  Population1    ST_LHS(6)         1    11    0      NA      NA      NA
#> 17  Population1    ST_LHS(6)         2    13    0      NA      NA      NA
#> 18  Population1    ST_LHS(6)         3    15    0      NA      NA      NA
#> 19  Population1    ST_LHS(7)         1    79   16      NA      NA      NA
#> 20  Population1    ST_LHS(7)         2    54   13      NA      NA      NA
#> 21  Population1    ST_LHS(7)         3    83    3      NA      NA      NA
#> 22  Population1    ST_LHS(8)         1    70   18      NA      NA      NA
#> 23  Population1    ST_LHS(8)         2    24    0      NA      NA      NA
#> 24  Population1    ST_LHS(8)         3    23    0      NA      NA      NA
#> 25  Population1    ST_LHS(9)         1    NA  382  0.9339  0.0628      28
#> 26  Population1    ST_LHS(9)         2    NA  722  0.9538  0.0402      42
#> 27  Population1    ST_LHS(9)         3    NA  386  0.9679  0.0181      46
#> 28  Population1   ST_LHS(10)         1    32    0      NA      NA      NA
#> 29  Population1   ST_LHS(10)         2    34    0      NA      NA      NA
#> 30  Population1   ST_LHS(10)         3    99   15      NA      NA      NA
#> 31  Population1   ST_LHS(11)         1    NA  785  0.9674  0.0217      62
#> 32  Population1   ST_LHS(11)         2    NA  814  0.9658  0.0319      64
#> 33  Population1   ST_LHS(11)         3    NA  809  0.9710  0.0334      71
#> 34  Population1   ST_LHS(12)         1    NA 1253  0.9864  0.0160     127
#> 35  Population1   ST_LHS(12)         2    NA 1330  0.9845  0.0180     125
#> 36  Population1   ST_LHS(12)         3    NA 1290  0.9861  0.0116     124
#> 37  Population1   ST_LHS(13)         1    33    0      NA      NA      NA
#> 38  Population1   ST_LHS(13)         2    36    0      NA      NA      NA
#> 39  Population1   ST_LHS(13)         3    24    0      NA      NA      NA
#> 40  Population1   ST_LHS(14)         1    NA  557  0.9709  0.0215      67
#> 41  Population1   ST_LHS(14)         2    NA  588  0.9736  0.0221      68
#> 42  Population1   ST_LHS(14)         3    NA  527  0.9732  0.0285      63
#> 43  Population1   ST_LHS(15)         1    NA  385  0.9575  0.0286      39
#> 44  Population1   ST_LHS(15)         2    NA  552  0.9549  0.0489      49
#> 45  Population1   ST_LHS(15)         3    NA  580  0.9635  0.0362      50
#> 46  Population1   ST_LHS(16)         1    11    0      NA      NA      NA
#> 47  Population1   ST_LHS(16)         2     8    0      NA      NA      NA
#> 48  Population1   ST_LHS(16)         3    11    0      NA      NA      NA
#> 49  Population1   ST_LHS(17)         1    NA 1166  0.9756  0.0232      77
#> 50  Population1   ST_LHS(17)         2    NA 1057  0.9755  0.0237      74
#> 51  Population1   ST_LHS(17)         3    NA 1098  0.9763  0.0246      80
#> 52  Population1   ST_LHS(18)         1    12    0      NA      NA      NA
#> 53  Population1   ST_LHS(18)         2     4    0      NA      NA      NA
#> 54  Population1   ST_LHS(18)         3     6    0      NA      NA      NA
#> 55  Population1   ST_LHS(19)         1     9    0      NA      NA      NA
#> 56  Population1   ST_LHS(19)         2     9    0      NA      NA      NA
#> 57  Population1   ST_LHS(19)         3    10    0      NA      NA      NA
#> 58  Population1   ST_LHS(20)         1     9    0      NA      NA      NA
#> 59  Population1   ST_LHS(20)         2     9    0      NA      NA      NA
#> 60  Population1   ST_LHS(20)         3    12    0      NA      NA      NA
#> 61  Population1   ST_LHS(21)         1    NA 1106  0.9642  0.0407      48
#> 62  Population1   ST_LHS(21)         2    NA 1271  0.9636  0.0393      47
#> 63  Population1   ST_LHS(21)         3    NA 1544  0.9632  0.0298      59
#> 64  Population1   ST_LHS(22)         1    NA  538  0.9716  0.0353      65
#> 65  Population1   ST_LHS(22)         2    NA  587  0.9713  0.0187      55
#> 66  Population1   ST_LHS(22)         3    NA  541  0.9664  0.0314      50
#> 67  Population1   ST_LHS(23)         1    36    0      NA      NA      NA
#> 68  Population1   ST_LHS(23)         2    19    0      NA      NA      NA
#> 69  Population1   ST_LHS(23)         3    30    0      NA      NA      NA
#> 70  Population1   ST_LHS(24)         1    NA   88  0.8917  0.0909      11
#> 71  Population1   ST_LHS(24)         2   101   12      NA      NA      NA
#> 72  Population1   ST_LHS(24)         3    NA  232  0.9211  0.0776      23
#> 73  Population1   ST_LHS(25)         1    NA 1087  0.9744  0.0313      69
#> 74  Population1   ST_LHS(25)         2    NA 1126  0.9703  0.0284      68
#> 75  Population1   ST_LHS(25)         3    NA 1218  0.9742  0.0263      62
#> 76  Population1   ST_LHS(26)         1    NA  438  0.9634  0.0137      42
#> 77  Population1   ST_LHS(26)         2    NA  390  0.9580  0.0359      41
#> 78  Population1   ST_LHS(26)         3    NA  418  0.9211  0.0622      29
#> 79  Population1   ST_LHS(27)         1    NA  465  0.9588  0.0409      42
#> 80  Population1   ST_LHS(27)         2    NA  465  0.9734  0.0129      60
#> 81  Population1   ST_LHS(27)         3    NA  437  0.9556  0.0526      47
#> 82  Population1   ST_LHS(28)         1    NA  464  0.9707  0.0302      55
#> 83  Population1   ST_LHS(28)         2    NA  244  0.9585  0.0492      38
#> 84  Population1   ST_LHS(28)         3    NA  353  0.9744  0.0283      61
#> 85  Population1   ST_LHS(29)         1    NA  408  0.9528  0.0515      47
#> 86  Population1   ST_LHS(29)         2    NA  468  0.9499  0.0769      44
#> 87  Population1   ST_LHS(29)         3    NA  419  0.9727  0.0167      61
#> 88  Population1   ST_LHS(30)         1    NA  461  0.9291  0.0651      23
#> 89  Population1   ST_LHS(30)         2    NA  291  0.8715  0.1375      14
#> 90  Population1   ST_LHS(30)         3    NA   79  0.9101  0.1013      16
#> 91  Population1   ST_LHS(31)         1    NA  930  0.9830  0.0183     112
#> 92  Population1   ST_LHS(31)         2    NA 1010  0.9845  0.0149     118
#> 93  Population1   ST_LHS(31)         3    NA  764  0.9861  0.0196     124
#> 94  Population1   ST_LHS(32)         1    NA  663  0.9808  0.0226      87
#> 95  Population1   ST_LHS(32)         2    NA  580  0.9728  0.0276      74
#> 96  Population1   ST_LHS(32)         3    NA  678  0.9803  0.0206      88
#> 97  Population1   ST_LHS(33)         1    NA  781  0.9715  0.0320      58
#> 98  Population1   ST_LHS(33)         2    NA  796  0.9645  0.0226      46
#> 99  Population1   ST_LHS(33)         3    NA  847  0.9447  0.0626      41
#> 100 Population1   ST_LHS(34)         1    NA  476  0.9742  0.0336      74
#> 101 Population1   ST_LHS(34)         2    NA  577  0.9751  0.0191      73
#> 102 Population1   ST_LHS(34)         3    NA  603  0.9792  0.0116      81
#> 103 Population1   ST_LHS(35)         1    NA  355  0.9282  0.0620      34
#> 104 Population1   ST_LHS(35)         2    NA  187  0.9366  0.0802      24
#> 105 Population1   ST_LHS(35)         3    NA  879  0.9445  0.0478      32
#> 106 Population1   ST_LHS(36)         1    12    0      NA      NA      NA
#> 107 Population1   ST_LHS(36)         2    18    0      NA      NA      NA
#> 108 Population1   ST_LHS(36)         3    13    0      NA      NA      NA
#> 109 Population1   ST_LHS(37)         1    NA  982  0.9855  0.0102     118
#> 110 Population1   ST_LHS(37)         2    NA  996  0.9838  0.0050     116
#> 111 Population1   ST_LHS(37)         3    NA  884  0.9837  0.0147     121
#> 112 Population1   ST_LHS(38)         1    15    0      NA      NA      NA
#> 113 Population1   ST_LHS(38)         2    24    0      NA      NA      NA
#> 114 Population1   ST_LHS(38)         3    21    0      NA      NA      NA
#> 115 Population1   ST_LHS(39)         1     7    0      NA      NA      NA
#> 116 Population1   ST_LHS(39)         2    10    0      NA      NA      NA
#> 117 Population1   ST_LHS(39)         3    10    0      NA      NA      NA
#> 118 Population1   ST_LHS(40)         1    NA 1466  0.9883  0.0109     149
#> 119 Population1   ST_LHS(40)         2    NA 1429  0.9880  0.0140     154
#> 120 Population1   ST_LHS(40)         3    NA 1464  0.9882  0.0075     160
#> 121 Population1   ST_LHS(41)         1    29    0      NA      NA      NA
#> 122 Population1   ST_LHS(41)         2    24    0      NA      NA      NA
#> 123 Population1   ST_LHS(41)         3    22    0      NA      NA      NA
#> 124 Population1   ST_LHS(42)         1    NA  755  0.9612  0.0477      49
#> 125 Population1   ST_LHS(42)         2    NA  731  0.9546  0.0451      34
#> 126 Population1   ST_LHS(42)         3    NA  776  0.9663  0.0245      61
#> 127 Population1   ST_LHS(43)         1    31    0      NA      NA      NA
#> 128 Population1   ST_LHS(43)         2    55    0      NA      NA      NA
#> 129 Population1   ST_LHS(43)         3    20    0      NA      NA      NA
#> 130 Population1   ST_LHS(44)         1    33    0      NA      NA      NA
#> 131 Population1   ST_LHS(44)         2    39    0      NA      NA      NA
#> 132 Population1   ST_LHS(44)         3    35   10      NA      NA      NA
#> 133 Population1   ST_LHS(45)         1    48    0      NA      NA      NA
#> 134 Population1   ST_LHS(45)         2    50    0      NA      NA      NA
#> 135 Population1   ST_LHS(45)         3    30    0      NA      NA      NA
#> 136 Population1   ST_LHS(46)         1    NA 1243  0.9709  0.0241      61
#> 137 Population1   ST_LHS(46)         2    NA 1136  0.9795  0.0194      89
#> 138 Population1   ST_LHS(46)         3    NA 1216  0.9774  0.0214      80
#> 139 Population1   ST_LHS(47)         1   109   23      NA      NA      NA
#> 140 Population1   ST_LHS(47)         2    NA   52  0.8600  0.0962      10
#> 141 Population1   ST_LHS(47)         3    NA  379  0.9333  0.0580      26
#> 142 Population1   ST_LHS(48)         1    NA  213  0.9194  0.0704      21
#> 143 Population1   ST_LHS(48)         2    NA  183  0.8876  0.1475      21
#> 144 Population1   ST_LHS(48)         3    NA   81  0.8112  0.2222      13
#> 145 Population1   ST_LHS(49)         1    NA  682  0.9502  0.0469      38
#> 146 Population1   ST_LHS(49)         2    NA  681  0.9550  0.0426      40
#> 147 Population1   ST_LHS(49)         3    NA  690  0.9524  0.0348      39
#> 148 Population1   ST_LHS(50)         1    NA 1197  0.9884  0.0109     157
#> 149 Population1   ST_LHS(50)         2    NA 1036  0.9887  0.0068     156
#> 150 Population1   ST_LHS(50)         3    NA  993  0.9875  0.0111     134
#> 151 Population1 ST_LHS(Base)         1     1    0      NA      NA      NA
#> 152 Population1 ST_LHS(Base)         2     1    0      NA      NA      NA
#> 153 Population1 ST_LHS(Base)         3     1    0      NA      NA      NA
#>