Collated results from Vortex scenarios - Campbell et al (2016)

A dataset with the results from Vortex scenarios used in Campbell et al (2016) to simulate major application of control measures in every 5 year cycle. Vortex outputs, from the project named 'Starlingv3PopBased' and the sensitivity test scenario 'MReductEvy5' (.stdat files), were collated with collate_dat.

Format

a data.frame with 1020 observations of 47 variables.

Source

Campbell et al. (2016). Assessing the economic benefits of starling detection and control to Western Australia. Australasian Journal of Environmental Management, 23, 81-99. DOI:10.1080/14486563.2015.1028486

Examples

data("sta.evy5")
head(sta.evy5)

#>        scen.name pop.name Year PExtant SE.PExtant. PExtinct SE.PExtinct.  r.det
#> 1 MReductEvy5(1)   Border    0       1           0        0            0 0.5247
#> 2 MReductEvy5(1)   Border    1       1           0        0            0 0.5247
#> 3 MReductEvy5(1)   Border    2       1           0        0            0 0.5247
#> 4 MReductEvy5(1)   Border    3       1           0        0            0 0.5247
#> 5 MReductEvy5(1)   Border    4       1           0        0            0 0.5247
#> 6 MReductEvy5(1)   Border    5       1           0        0            0 0.5247
#>   r.stoch  SE.r.  SD.r. Nextant SE.Nextant. SD.Nextant.    Nall SE.Nall.
#> 1  0.0000 0.0000 0.0000  632.00        0.00        0.00  632.00     0.00
#> 2  0.4122 0.0043 0.4281 1032.61        3.71      371.08 1032.61     0.04
#> 3  0.4051 0.0043 0.4312 1671.63        8.75      874.63 1671.63     0.09
#> 4  0.4009 0.0044 0.4389 2712.70       18.24     1823.50 2712.70     0.18
#> 5  0.3970 0.0045 0.4478 4408.86       36.14     3614.24 4408.86     0.36
#> 6 -0.1507 0.0070 0.7015 4590.36       49.40     4939.92 4590.36     0.49
#>   SD.Nall. Het SE.Het. SD.Het. Inbr SE.Inbr. SD.Inbr. Nalleles SE.Nalleles.
#> 1     0.00   0       0       0    1        0        0        0            0
#> 2   371.08   0       0       0    1        0        0        0            0
#> 3   874.63   0       0       0    1        0        0        0            0
#> 4  1823.50   0       0       0    1        0        0        0            0
#> 5  3614.24   0       0       0    1        0        0        0            0
#> 6  4939.92   0       0       0    1        0        0        0            0
#>   SD.Nalleles. Nlethals SE.Nlethals. SD.Nlethals. TE SE.TE. SD.TE. medTE NRANDx
#> 1            0        0            0            0  0      0      0     0 0.0017
#> 2            0        0            0            0  0      0      0     0 0.0040
#> 3            0        0            0            0  0      0      0     0 0.0123
#> 4            0        0            0            0  0      0      0     0 0.0043
#> 5            0        0            0            0  0      0      0     0 0.0157
#> 6            0        0            0            0  0      0      0     0 0.0017
#>   SE.NRANDx. SD.NRANDx. immigBord SE.immigBord. SD.immigBord. immigSC
#> 1      1e-04     1.0068   14.4194         6e-04        6.3242  2.4943
#> 2      1e-04     0.9879   14.5315         6e-04        6.3326  2.4903
#> 3      1e-04     0.9971   14.4120         6e-04        6.3328  2.5074
#> 4      1e-04     0.9990   14.5192         6e-04        6.3866  2.5237
#> 5      1e-04     1.0054   14.5297         6e-04        6.4000  2.4942
#> 6      1e-04     1.0033   14.4552         6e-04        6.3757  2.4978
#>   SE.immigSC. SD.immigSC. SV1 SE.SV1. SD.SV1. SD.PExtant. SD.PExtinct.
#> 1       1e-04      1.1095   1       0       0           0            0
#> 2       1e-04      1.1119   1       0       0           0            0
#> 3       1e-04      1.1191   1       0       0           0            0
#> 4       1e-04      1.1255   1       0       0           0            0
#> 5       1e-04      1.1219   1       0       0           0            0
#> 6       1e-04      1.1138   1       0       0           0            0