Collated results from Vortex scenarios - Campbell et al (2016)

A dataset with the results from Vortex scenarios used in Campbell et al (2016) to simulate major application of control measures in every 5 year cycle, maintaining 2011 levels of investment. Vortex outputs, from the project named 'Starlingv3PopBased' and the sensitivity test scenario 'MReduction_B11_09Evy5' (.stdat files), were collated with collate_dat.

Format

a data.frame with 1020 observations of 47 variables.

Source

Campbell et al. (2016). Assessing the economic benefits of starling detection and control to Western Australia. Australasian Journal of Environmental Management, 23, 81-99. DOI:10.1080/14486563.2015.1028486

Examples

data("sta.evy5.b11")
head(sta.evy5.b11)

#>                  scen.name pop.name Year PExtant SE.PExtant. PExtinct
#> 1 MReduction_B11_09Evy5(1)   Border    0       1           0        0
#> 2 MReduction_B11_09Evy5(1)   Border    1       1           0        0
#> 3 MReduction_B11_09Evy5(1)   Border    2       1           0        0
#> 4 MReduction_B11_09Evy5(1)   Border    3       1           0        0
#> 5 MReduction_B11_09Evy5(1)   Border    4       1           0        0
#> 6 MReduction_B11_09Evy5(1)   Border    5       1           0        0
#>   SE.PExtinct.  r.det r.stoch  SE.r.  SD.r. Nextant SE.Nextant. SD.Nextant.
#> 1            0 0.2668  0.0000 0.0000 0.0000  632.00        0.00        0.00
#> 2            0 0.2668  0.1130 0.0057 0.5724  810.01        3.70      369.56
#> 3            0 0.2668  0.1012 0.0058 0.5795 1028.02        7.00      699.66
#> 4            0 0.2668  0.0926 0.0059 0.5872 1283.09       11.23     1122.90
#> 5            0 0.2668  0.1076 0.0057 0.5727 1618.19       17.15     1714.54
#> 6            0 0.2668 -0.1259 0.0069 0.6883 1709.58       22.93     2293.49
#>      Nall SE.Nall. SD.Nall. Het SE.Het. SD.Het. Inbr SE.Inbr. SD.Inbr. Nalleles
#> 1  632.00     0.00     0.00   0       0       0    1        0        0        0
#> 2  810.01     0.04   369.56   0       0       0    1        0        0        0
#> 3 1028.02     0.07   699.66   0       0       0    1        0        0        0
#> 4 1283.09     0.11  1122.90   0       0       0    1        0        0        0
#> 5 1618.19     0.17  1714.54   0       0       0    1        0        0        0
#> 6 1709.58     0.23  2293.49   0       0       0    1        0        0        0
#>   SE.Nalleles. SD.Nalleles. Nlethals SE.Nlethals. SD.Nlethals. TE SE.TE. SD.TE.
#> 1            0            0        0            0            0  0      0      0
#> 2            0            0        0            0            0  0      0      0
#> 3            0            0        0            0            0  0      0      0
#> 4            0            0        0            0            0  0      0      0
#> 5            0            0        0            0            0  0      0      0
#> 6            0            0        0            0            0  0      0      0
#>   medTE  NRANDx SE.NRANDx. SD.NRANDx. immigBord SE.immigBord. SD.immigBord.
#> 1     0 -0.0011      1e-04     1.0102   14.4091         6e-04        6.3782
#> 2     0 -0.0040      1e-04     1.0026   14.5168         6e-04        6.3307
#> 3     0 -0.0054      1e-04     1.0037   14.5307         6e-04        6.3107
#> 4     0 -0.0017      1e-04     1.0048   14.4816         6e-04        6.3436
#> 5     0  0.0141      1e-04     0.9956   14.4873         6e-04        6.3568
#> 6     0  0.0130      1e-04     1.0044   14.3948         6e-04        6.2980
#>   immigSC SE.immigSC. SD.immigSC. SV1 SE.SV1. SD.SV1. SD.PExtant. SD.PExtinct.
#> 1  2.5066       1e-04      1.1206   1       0       0           0            0
#> 2  2.4983       1e-04      1.1192   1       0       0           0            0
#> 3  2.4777       1e-04      1.1163   1       0       0           0            0
#> 4  2.4979       1e-04      1.1148   1       0       0           0            0
#> 5  2.5016       1e-04      1.1171   1       0       0           0            0
#> 6  2.5119       1e-04      1.1215   1       0       0           0            0