Collate Vortex .dat or .stdat output files into one data.frame

collate_dat collates all Vortex output files matching a given project name (and scenario name when relevant) in a given directory into one data.frame using collate_one_dat.

collate_dat(project, runs, scenario = NULL, dir_in = NULL,
  save2disk = TRUE, dir_out = "ProcessedData", verbose = TRUE)

Arguments

project
The Vortex project name to be imported

runs
The number of Vortex simulation runs

scenario
The scenario name if ST, default: NULL

dir_in
The local folder containing Vortex files, default: NULL. If not specified, will fall back to use current working directory.

save2disk
Whether to save the data as rda and csv, default: TRUE

dir_out
The local path to store the output. Default: ProcessedData

verbose
Progress messages, default: TRUE

Value

a data.frame with data from all matching Vortex files or NULL

Details

The number of Vortex simulation runs has to be specified, as it cannot be inferred by vortexR.

To read Vortex output files from Sensitivity Testing with the extension '.stdat', specify the scenario name. If scenario=NULL, all files with extension .dat, matching the project name will be imported. dir_in may contain other files; only files matching the project (and, optionally, the scenario) name will be read. dir_out is created within the working directory unless a full path is provided.

If no matching files are found in the given directory, an error is reported.

When verbose=TRUE the progress (i.e. the file being read) is reported on screen.

Examples

# Using Campbell et al. and Pacioni et al. example files.
# See ?pacioni and ?campbell for more details on example files.
require(vortexRdata)
camp.dir <- system.file('extdata', 'campbell', package='vortexRdata')
pac.dir <- system.file('extdata', 'pacioni', package='vortexRdata')

# Campbell example, project 'Starlingv3PopBased' (.dat)
starling <- collate_dat('Starlingv3PopBased', 10000,
            dir_in=camp.dir, save2disk=FALSE)

#> INFO vortexR::collate_dat found 8 matching files in /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/campbell:

#> vortexR::collate_dat is parsing:

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/campbell/Starlingv3PopBased_baseline.dat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/campbell/Starlingv3PopBased_Control_09.dat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/campbell/Starlingv3PopBased_Control_09Tech.dat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/campbell/Starlingv3PopBased_Control_11.dat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/campbell/Starlingv3PopBased_Control_11Tech.dat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/campbell/Starlingv3PopBased_Control_Border11.dat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/campbell/Starlingv3PopBased_Control_D09.dat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/campbell/Starlingv3PopBased_Control_SC11.dat

# Read data from all .stdat of the project 'Pacioni_et_al' and the ST scenario
#   'ST_Classic' and store the output in the object 'woylie.st.classic'
woylie.st.classic <- collate_dat('Pacioni_et_al', 3, scenario = 'ST_Classic',
                     dir_in = pac.dir, save2disk=FALSE)

#> INFO vortexR::collate_dat found 24 matching files in /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni:

#> vortexR::collate_dat is parsing:

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(1).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(2).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(3).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(4).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(5).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(6).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(7).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(8).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(9).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(10).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(11).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(12).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(13).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(14).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(15).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(16).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(17).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(18).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(19).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(20).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(21).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(22).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(23).stdat

#> /home/CORPORATEICT/florianm/R/x86_64-pc-linux-gnu-library/3.3/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_Classic(Base).stdat

# Save collated data as .Rda and .txt

# NOT RUN {
# Read data from all .stdat of the project 'Pacioni_et_al' and the ST scenario
#   'ST_Classic'. Store the output in the object 'woylie.st.classic' and save
#   to disk
woylie.st.classic <- collate_dat('Pacioni_et_al', 3, scenario = 'ST_Classic',
                     dir_in = pac.dir, save2disk=TRUE)
# }