Collate Vortex .run output files

collate_run collates all Vortex output files with extension .run matching the project and scenario name in a given directory into one named list.

collate_run(
  project,
  scenario,
  npops = 1,
  dir_in = NULL,
  save2disk = TRUE,
  dir_out = "ProcessedData",
  verbose = TRUE
)

Arguments

project
The Vortex project name to be imported

scenario
The scenario name

npops
The total number of simulated populations including the metapopulation

dir_in
The local folder containing Vortex files, default: NULL. If not specified, will fall back to use current working directory.

save2disk
Whether to save the data as rda and csv, default: TRUE

dir_out
The local path to store the output. Default: ProcessedData

verbose
Progress messages, default: TRUE

Value

a list with two elements: run, a data.frame with data from all Vortex files and lrun, where the same data are re-arranged in long format

Details

dir_in may contain other files; only files matching the project and the scenario name will be read.

dir_out is created within the working directory unless a full path is provided.

If no matching files are found in the given directory, an error is reported.

When verbose=TRUE the progress (i.e. the file being read) is reported on screen.

Examples

# Using Pacioni et al. example files. See ?pacioni for more details.
pac.dir <- system.file('extdata', 'pacioni', package='vortexRdata')
# Run collate_run on all .run of the project 'Pacioni_et_al' and
run <- collate_run('Pacioni_et_al', 'ST_LHS', 1, dir_in=pac.dir,
                   save2disk=FALSE)

#> INFO vortexR::collate_run found 51 matching files in C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni:

#> vortexR::collate_run is parsing:

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(1).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(2).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(3).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(4).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(5).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(6).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(7).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(8).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(9).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(10).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(11).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(12).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(13).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(14).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(15).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(16).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(17).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(18).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(19).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(20).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(21).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(22).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(23).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(24).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(25).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(26).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(27).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(28).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(29).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(30).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(31).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(32).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(33).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(34).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(35).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(36).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(37).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(38).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(39).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(40).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(41).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(42).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(43).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(44).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(45).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(46).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(47).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(48).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(49).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(50).run

#> C:/Users/Carlo/Documents/R/win-library/3.6/vortexRdata/extdata/pacioni/Pacioni_et_al_ST_LHS(Base).run