Collated results from Vortex scenarios - Pacioni et al. (2017)

Data from Pacioni et al. (2017) used to conduct a sensitivity analysis on demographic parameters. Vortex outputs, from the project named 'Pacioni_et_al' and (Single-Factor) sensitivity test scenario 'ST_Classic' (.yr files), were collated with collate_yr.

Format

A named list of two elements: all (8712 obs, 26 var), means (2904 obs, 25 var).

Source

Pacioni, C., Williams, M., Lacy RC, Spencer, P.B.S. and Wayne, A.F. (2017) Predators and genetic fitness: key threatening factors for the conservation of bettong species. Pacific Conservation Biology. DOI:10.1071/PC17002

Examples

data("pac.yr")
head(pac.yr)

#> $census
#>               Scenario Iteration Year Npop1 AMpop1 AFpop1 Subadultspop1 Juvpop1
#>    1:    ST_Classic(1)         1    0   300     97    117            24      62
#>    2:    ST_Classic(1)         2    0   300     97    117            24      62
#>    3:    ST_Classic(1)         3    0   300     97    117            24      62
#>    4:    ST_Classic(1)         1    1   361    116    137            28      80
#>    5:    ST_Classic(1)         2    1   357    115    136            29      77
#>   ---                                                                          
#> 8708: ST_Classic(Base)         2  119  2081    712    798           169     402
#> 8709: ST_Classic(Base)         3  119  1887    614    768           167     338
#> 8710: ST_Classic(Base)         1  120  1775    630    680            97     368
#> 8711: ST_Classic(Base)         2  120  2110    832    944           165     169
#> 8712: ST_Classic(Base)         3  120  1869    626    755           122     366
#>       nDamspop1 nBroodspop1 nProgenypop1 nImmigrantspop1 nEmigrantspop1
#>    1:         0           0            0               0              0
#>    2:         0           0            0               0              0
#>    3:         0           0            0               0              0
#>    4:        81          81           81               0              0
#>    5:        80          80           80               0              0
#>   ---                                                                  
#> 8708:       419         419          419               0              0
#> 8709:       406         406          406               0              0
#> 8710:       464         464          464               0              0
#> 8711:       180         180          180               0              0
#> 8712:       442         442          442               0              0
#>       nHarvestedpop1 nSupplementedpop1 GS1pop1 GS2pop1 GS3pop1 GS4pop1 GS5pop1
#>    1:              0                 0     500       1       1       1     0.1
#>    2:              0                 0     500       1       1       1     0.1
#>    3:              0                 0     500       1       1       1     0.1
#>    4:              0                 0     500       1       1       1     0.1
#>    5:              0                 0     500       1       1       1     0.1
#>   ---                                                                         
#> 8708:              0                 0    2000       1       1       1     0.1
#> 8709:              0                 0    2000       1       1       1     0.1
#> 8710:              0                 0    2000       1       1       1     0.1
#> 8711:              0                 0    2000       1       1       1     0.1
#> 8712:              0                 0    2000       1       1       1     0.1
#>       GS6pop1 GS7pop1 PS1pop1 PS2pop1 PS3pop1 PS4pop1
#>    1:      80     300   0.000   0.199     4.4   0.436
#>    2:      80     300   0.000   0.202     4.4   0.635
#>    3:      80     300   0.000   0.198     4.4   0.348
#>    4:      80     300   0.000   0.198     4.4   0.516
#>    5:      80     300   0.000   0.199     4.4   0.467
#>   ---                                                
#> 8708:      80     300   0.008   0.201     4.4   0.505
#> 8709:      80     300   0.008   0.200     4.4   0.344
#> 8710:      80     300   0.008   0.199     4.4   0.789
#> 8711:      80     300   0.008   0.200     4.4   0.403
#> 8712:      80     300   0.008   0.200     4.4   0.769
#> 
#> $census_means
#>               Scenario Year     Npop1   AMpop1   AFpop1 Subadultspop1   Juvpop1
#>    1:    ST_Classic(1)    0  300.0000  97.0000 117.0000      24.00000  62.00000
#>    2:    ST_Classic(1)    1  362.0000 117.0000 138.0000      28.66667  78.33333
#>    3:    ST_Classic(1)    2  433.6667 141.0000 172.3333      34.00000  86.33333
#>    4:    ST_Classic(1)    3  500.6667 165.0000 200.0000      35.66667 100.00000
#>    5:    ST_Classic(1)    4  481.3333 166.6667 202.6667      40.66667  71.33333
#>   ---                                                                          
#> 2900: ST_Classic(Base)  116 2051.3333 832.6667 835.0000      93.66667 290.00000
#> 2901: ST_Classic(Base)  117 1858.6667 760.6667 794.0000     121.00000 183.00000
#> 2902: ST_Classic(Base)  118 1914.0000 715.0000 729.0000      74.33333 395.66667
#> 2903: ST_Classic(Base)  119 1920.6667 665.3333 769.0000     159.00000 327.33333
#> 2904: ST_Classic(Base)  120 1918.0000 696.0000 793.0000     128.00000 301.00000
#>       nDamspop1 nBroodspop1 nProgenypop1 nImmigrantspop1 nEmigrantspop1
#>    1:   0.00000     0.00000      0.00000               0              0
#>    2:  81.00000    81.00000     81.00000               0              0
#>    3:  90.33333    90.33333     90.33333               0              0
#>    4: 104.33333   104.33333    104.33333               0              0
#>    5:  87.00000    87.00000     87.00000               0              0
#>   ---                                                                  
#> 2900: 314.66667   314.66667    314.66667               0              0
#> 2901: 220.66667   220.66667    220.66667               0              0
#> 2902: 484.33333   484.33333    484.33333               0              0
#> 2903: 377.00000   377.00000    377.00000               0              0
#> 2904: 362.00000   362.00000    362.00000               0              0
#>       nHarvestedpop1 nSupplementedpop1 GS1pop1 GS2pop1 GS3pop1 GS4pop1 GS5pop1
#>    1:              0                 0     500       1       1       1     0.1
#>    2:              0                 0     500       1       1       1     0.1
#>    3:              0                 0     500       1       1       1     0.1
#>    4:              0                 0     500       1       1       1     0.1
#>    5:              0                 0     500       1       1       1     0.1
#>   ---                                                                         
#> 2900:              0                 0    2000       1       1       1     0.1
#> 2901:              0                 0    2000       1       1       1     0.1
#> 2902:              0                 0    2000       1       1       1     0.1
#> 2903:              0                 0    2000       1       1       1     0.1
#> 2904:              0                 0    2000       1       1       1     0.1
#>       GS6pop1 GS7pop1      PS1pop1   PS2pop1 PS3pop1   PS4pop1
#>    1:      80     300 0.0000000000 0.1996667     4.4 0.4730000
#>    2:      80     300 0.0000000000 0.1993333     4.4 0.6406667
#>    3:      80     300 0.0000000000 0.1983333     4.4 0.2543333
#>    4:      80     300 0.0000000000 0.1980000     4.4 0.3226667
#>    5:      80     300 0.0003333333 0.2003333     4.4 0.4483333
#>   ---                                                         
#> 2900:      80     300 0.0076666667 0.2003333     4.4 0.7293333
#> 2901:      80     300 0.0080000000 0.2000000     4.4 0.4293333
#> 2902:      80     300 0.0080000000 0.2000000     4.4 0.4790000
#> 2903:      80     300 0.0080000000 0.2000000     4.4 0.5603333
#> 2904:      80     300 0.0080000000 0.1996667     4.4 0.6536667
#>