Collated results from Vortex scenarios - Campbell et al (2016)

A dataset with the results from the main Vortex scenarios used in Campbell et al (2016). Vortex outputs, from the project named 'Starlingv3PopBased' (.dat files), were collated with collate_dat.

Format

a data.frame with 1632 observations of 44 variables.

Source

Campbell et al. (2016). Assessing the economic benefits of starling detection and control to Western Australia. Australasian Journal of Environmental Management, 23, 81-99. DOI:10.1080/14486563.2015.1028486

Examples

data("sta.main")
head(sta.main)

#>   scen.name pop.name Year PExtant SE.PExtant. PExtinct SE.PExtinct.  r.det
#> 1  baseline   Border    0       1           0        0            0 0.5247
#> 2  baseline   Border    1       1           0        0            0 0.5247
#> 3  baseline   Border    2       1           0        0            0 0.5247
#> 4  baseline   Border    3       1           0        0            0 0.5247
#> 5  baseline   Border    4       1           0        0            0 0.5247
#> 6  baseline   Border    5       1           0        0            0 0.5247
#>   r.stoch  SE.r.  SD.r. Nextant SE.Nextant. SD.Nextant.    Nall SE.Nall.
#> 1  0.0000 0.0000 0.0000  632.00        0.00        0.00  632.00     0.00
#> 2  0.4131 0.0043 0.4334 1035.31        3.76      375.71 1035.31     0.04
#> 3  0.4085 0.0044 0.4364 1686.58        8.85      884.59 1686.58     0.09
#> 4  0.4049 0.0044 0.4418 2736.53       18.01     1801.39 2736.53     0.18
#> 5  0.4009 0.0044 0.4424 4426.31       34.76     3475.57 4426.31     0.35
#> 6  0.4050 0.0044 0.4419 7180.73       65.27     6527.49 7180.73     0.65
#>   SD.Nall. Het SE.Het. SD.Het. Inbr SE.Inbr. SD.Inbr. Nalleles SE.Nalleles.
#> 1     0.00   0       0       0    1        0        0        0            0
#> 2   375.71   0       0       0    1        0        0        0            0
#> 3   884.59   0       0       0    1        0        0        0            0
#> 4  1801.39   0       0       0    1        0        0        0            0
#> 5  3475.57   0       0       0    1        0        0        0            0
#> 6  6527.49   0       0       0    1        0        0        0            0
#>   SD.Nalleles. Nlethals SE.Nlethals. SD.Nlethals. TE SE.TE. SD.TE. medTE
#> 1            0        0            0            0  0      0      0     0
#> 2            0        0            0            0  0      0      0     0
#> 3            0        0            0            0  0      0      0     0
#> 4            0        0            0            0  0      0      0     0
#> 5            0        0            0            0  0      0      0     0
#> 6            0        0            0            0  0      0      0     0
#>    NRANDx SE.NRANDx. SD.NRANDx. immigBord SE.immigBord. SD.immigBord. immigSC
#> 1  0.0033      1e-04     1.0038   14.5338         6e-04        6.3540  2.4959
#> 2 -0.0078      1e-04     1.0029   14.4827         6e-04        6.3426  2.4926
#> 3  0.0012      1e-04     1.0128   14.5746         6e-04        6.3438  2.4864
#> 4  0.0084      1e-04     0.9994   14.3983         6e-04        6.3546  2.4937
#> 5 -0.0113      1e-04     1.0065   14.3340         6e-04        6.3282  2.4901
#> 6  0.0087      1e-04     1.0144   14.4516         6e-04        6.3690  2.4872
#>   SE.immigSC. SD.immigSC. SD.PExtant. SD.PExtinct.
#> 1       1e-04      1.1218           0            0
#> 2       1e-04      1.1130           0            0
#> 3       1e-04      1.1202           0            0
#> 4       1e-04      1.1318           0            0
#> 5       1e-04      1.1214           0            0
#> 6       1e-04      1.1233           0            0